+++ app/[lang]/(app)/(home)/(tablet)/restaurant/new-order/components/order-invoice/OrderInvoice.tsx
@@ -85,6 +85,7 @@ export default function OrderInvoice({ dic }: { dic: NewOrderDictionary }) {
    onPayment,
    onPaymentPcPos,
   },
+  access,
  } = useOrderBaseConfigContext();
  const { format } = useCurrencyFormatter();