+++ app/[lang]/(app)/(home)/(tablet)/restaurant/new-order/services/order-tools/OrderBaseConfigProvider.tsx
@@ -457,6 +457,7 @@ export default function OrderBaseConfigProvider({
  }, [systemPricing, orderInfoForm]);
 
  function onCloseOrder() {
+  if (!userAccessibility['restaurant']['order']['close']) return;
   setShowCloseOrder(true);
  }
  const { mutate: handleConfirmCloseOrder, isPending: isPendingCloseOrder } =
@@ -636,7 +637,11 @@ export default function OrderBaseConfigProvider({
   };
  }
 
+ const orderEditAccess = !!userOrder?.id
+  ? userAccessibility['restaurant']['order']['edit']
+  : userAccessibility['restaurant']['order']['add'];
  async function handleSaveOrder() {
+  if (!orderEditAccess) return;
   const orderInfoRes = await validateOrderInfo();
   if (!orderInfoRes) return;
   confirmSaveOrder({
@@ -700,6 +705,7 @@ export default function OrderBaseConfigProvider({
   });
 
  async function handlePayment(paymentData: OrderInvoicePayment) {
+  if (!userAccessibility['restaurant']['order']['payment']) return;
   const newOrderRes = await validateOrderInfo();
   if (!newOrderRes) return;
   handleConfirmPayment({
@@ -946,6 +952,13 @@ export default function OrderBaseConfigProvider({
    isError: systemPricingIsError,
    isLoading: systemPricingIsLoading,
    handleSetSystemPricing: onSetSystemPricing,
+  },
+  access: {
+   order: {
+    ...userAccessibility['restaurant']['order'],
+    edit: orderEditAccess,
+   },
+   shopItem: userAccessibility['restaurant']['orderItem'],
   },
  };